pythonocc-test/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aom=3.4.0=h0e60522_1
bzip2=1.0.8=h8ffe710_4
ca-certificates=2022.6.15=h5b45459_0
colorama=0.4.5=pyhd8ed1ab_0
curl=7.83.1=h789b8ee_0
double-conversion=3.2.0=h0e60522_0
eigen=3.4.0=h2d74725_0
expat=2.4.8=h39d44d4_0
ffmpeg=4.4.2=gpl_hbb6702b_107
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.0=hce3cb01_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
freeimage=3.18.0=h064cf08_10
freetype=2.12.1=h546665d_0
gettext=0.19.8.1=ha2e2712_1008
gl2ps=1.4.2=h0597ee9_0
glew=2.1.0=h39d44d4_2
glib=2.72.1=h7755175_0
glib-tools=2.72.1=h7755175_0
gst-plugins-base=1.20.3=he07aa86_0
gstreamer=1.20.3=hdff456e_0
hdf4=4.2.15=h0e5069d_4
hdf5=1.12.2=nompi_h2a0e4a3_100
icu=70.1=h0e60522_0
imath=3.1.5=h12d4b20_0
intel-openmp=2022.1.0=h57928b3_3787
jpeg=9e=h8ffe710_2
jsoncpp=1.9.5=h2d74725_1
jxrlib=1.1=h8ffe710_2
krb5=1.19.3=h1176d77_0
lcms2=2.12=h2a16943_0
lerc=4.0.0=h63175ca_0
libblas=3.9.0=16_win64_mkl
libcblas=3.9.0=16_win64_mkl
libclang=14.0.6=default_h77d9078_0
libclang13=14.0.6=default_h77d9078_0
libcurl=7.83.1=h789b8ee_0
libdeflate=1.13=h8ffe710_0
libffi=3.4.2=h8ffe710_5
libglib=2.72.1=h3be07f2_0
libiconv=1.16=he774522_0
liblapack=3.9.0=16_win64_mkl
libnetcdf=4.8.1=nompi_h85765be_104
libogg=1.3.4=h8ffe710_1
libpng=1.6.37=h1d00b33_4
libraw=0.20.2=hee1bdec_1
libsqlite=3.39.2=h8ffe710_1
libssh2=1.10.0=h680486a_3
libtheora=1.1.1=h8d14728_1005
libtiff=4.4.0=h92677e6_3
libvorbis=1.3.7=h0e60522_0
libwebp-base=1.2.4=h8ffe710_0
libxml2=2.9.14=hf5bbc77_4
libzip=1.9.2=hfed4ece_1
libzlib=1.2.12=h8ffe710_2
loguru=0.6.0=py310h5588dad_1
lz4-c=1.9.3=h8ffe710_1
mkl=2022.1.0=h6a75c08_874
numpy=1.23.2=py310h8a5b91a_0
occt=7.6.2=h8294837_3
openexr=3.1.5=hab3b255_0
openh264=2.3.0=h0e60522_0
openjpeg=2.5.0=hc9384bd_1
openssl=1.1.1q=h8ffe710_0
pcre=8.45=h0e60522_0
pip=22.2.2=pyhd8ed1ab_0
proj=9.0.1=h1cfcee9_1
pugixml=1.11.4=h0e60522_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt5=5.15.7=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.11.0=pypi_0
pyqt5designer=5.14.1=pypi_0
python=3.10.6=h9a09f29_0_cpython
python_abi=3.10=2_cp310
pythonocc-core=7.6.2=py310hf04ff9d_0
qt-main=5.15.4=h467ea89_2
rapidjson=1.1.0=ha925a31_1002
setuptools=65.2.0=py310h5588dad_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.39.2=h8ffe710_1
svgwrite=1.4.3=pyhd8ed1ab_0
svt-av1=1.2.1=h0e60522_0
tbb=2021.5.0=h2d74725_1
tbb-devel=2021.5.0=h2d74725_1
tk=8.6.12=h8ffe710_0
tzdata=2022c=h191b570_0
ucrt=10.0.20348.0=h57928b3_0
utfcpp=3.2.1=h57928b3_0
vc=14.2=hb210afc_6
vs2015_runtime=14.29.30037=h902a5da_6
vtk=9.1.0=qt_py310hbebbbde_214
wheel=0.37.1=pyhd8ed1ab_0
win32_setctime=1.1.0=pyhd8ed1ab_0
x264=1!164.3095=h8ffe710_2
x265=3.5=h2d74725_3
xz=5.2.6=h8d14728_0
zlib=1.2.12=h8ffe710_2
zstd=1.5.2=h7755175_4

pandas~=1.4.4
openpyxl~=3.0.10
networkx~=2.8.4